requirements.txt
langchain
langchain-core
langchain-community
langchain-groq
pytest
pytest-cov
python-dotenv
rich
nameko==3.0.0-rc11
